refactor: rearrange the intl example
